+++ sdk/search/Azure.Search.Documents/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 11.1.0-preview.1 (Unreleased)
 
+### Removed
+
+- Removed `$select` from the query parameters logged by default.  You can add it back via `SearchClientOptions.Diagnostics.LoggedQueryParameters("$select");` if desired.
 
 ## 11.0.0 (2020-07-07)