+++ sdk/storage/Azure.Storage.Files.DataLake/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Release History
 
-## 12.3.0-preview.2 (Unreleased)
+## 12.3.0-preview.3 (Unreleased)
+- Fixed bug where DataLakeFileSystemClient.SetAccessPolicy() sends DateTimeOffset.MinValue when StartsOn and ExpiresOn when not set in DataLakeAccessPolicy
+- Added nullable properties, PolicyStartsOn and PolicyExpiresOn to DataLakeAccessPolicy
+
+## 12.3.0-preview.2 (2020-07-27)
 - Fixed bug where DataLakeUriBuilder would return LastDirectoryOrFileName and DirectoryOrFilePath URL-encoded.
 - Updated DataLakeSasBuilder to correctly order raw string permissions and make the permissions lowercase.
 - Fixed bug where DataLakeFileClient.Query() failed when query response was > ~200 MB.