+++ sdk/textanalytics/Azure.AI.TextAnalytics/CHANGELOG.md
@@ -1,9 +1,15 @@
 # Release History
 
-## 1.1.0-preview.1 (Unreleased)
+## 5.1.0-preview.1 (Unreleased)
 ### Breaking changes
 - It defaults to the latest supported API version, which currently is `3.1-preview.1`.
 Note that new functionality hasn't been implemented in the client library.
+
+### New Features
+- `ErrorCode` value returned from the service is now surfaced in `RequestFailedException`.
+
+## 5.0.0 (2020-07-27)
+- Re-release of version `1.0.1` with updated version `5.0.0`.
 
 ## 1.0.1 (2020-06-23)